vision-service/requirements.txt
# Dependencias del microservicio de visión
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
opencv-python>=4.8.1.78
mediapipe>=0.10.13
numpy>=1.24.3
pillow>=10.1.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
loguru>=0.7.2
aiofiles>=23.2.1
httpx>=0.25.2
scipy>=1.11.4
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.0
